+++ src/main/java/chess/domain/chessBoard/ChessSpaceGenerator.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.chessBoard;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.EmptyPiece;
