+++ src/main/java/chess/domain/chessBoard/SpaceGenerator.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.chessBoard;
 
 import java.util.List;
