+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,9 +1,7 @@
 package chess.domain.piece;
 
-import chess.domain.Rule;
 import chess.domain.position.Difference;
 import chess.domain.position.Position;
-import chess.domain.position.PositionDifference;
 
 public class Bishop extends Piece {
