+++ src/main/java/chess/domain/piece/King.java
@@ -1,8 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.Rule;
 import chess.domain.position.Position;
-import chess.domain.position.PositionDifference;
 
 public class King extends Piece {
