+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,6 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Rule;
 import chess.domain.position.Position;
 import chess.domain.position.PositionDifference;
