+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,6 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Rule;
 import chess.domain.position.FileDifference;
 import chess.domain.position.Position;
 import chess.domain.position.RankDifference;
