+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,9 +1,7 @@
 package chess.domain.piece;
 
-import chess.domain.Rule;
 import chess.domain.position.Position;
 import chess.domain.position.FileDifference;
-import chess.domain.position.PositionDifference;
 import chess.domain.position.RankDifference;
 
 public class Rook extends Piece {
