+++ src/main/java/chess/domain/piece/Rule.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 import chess.domain.position.FileDifference;
 import chess.domain.position.RankDifference;
