+++ src/main/java/chess/view/PieceSign.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.view;
 
 import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
